mcp-weather/pyproject.toml
[project]
name = "weather-mcp-server"
version = "0.1.0"
description = "MCP server for receiving real weather data via the Open-Meteo API"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastmcp>=0.4.0",
    "uvicorn>=0.24.0",
    "starlette>=0.27.0",
    "python-dateutil>=2.8.2",
    "httpx>=0.25.0",
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
] 